Update InfraDeploymentGuide.md
Change of order for ## Credentials

By default, the template doesn’t have any Azure Account credentials preloaded. Hence, the first step to start showing subscriptions data is to sign-in with the right user credentials.

**IMPORTANT NOTE**: Power BI Desktop caches the credentials after the first logon. It is important to clear the credentials from Power BI desktop if you plan to switch between Azure Global and any other region like US Government or China. The same concept applies if you plan to switch between tenants. Otherwise, the staged credentials will be used again for the different Azure environments and the authentication or data load process will fail. 

As the clean of credential is a step that must be done. In the note it can be interpreted as you can choose not to do this step. By changing the order of the text, it will be more clear that this step is needed